fixed build problems

+++ zhuravlev.ivan/I3/finder_same_elements_in_two_strings.cpp
@@ -20,12 +20,12 @@ void finderSameElementsInTwoStrings(const char *first_string, const char *second
 
   char result_string[cnt]{};
 
-  for (int i = 0; i < cnt; ++i)
+  for (size_t i = 0; i < cnt; ++i)
   {
     result_string[i] = tmp_string[i];
   }
 
-  for (int i = 0; i <= cnt; ++i)
+  for (size_t i = 0; i <= cnt; ++i)
   {
     std::cout << result_string[i] << " ";
   }
